Solved CodeFactor issue
