+++ src/Notepads/Utilities/SessionUtility.cs
@@ -97,24 +97,4 @@
             }
         }
     }
-
-    internal static class BackupSessionUtility
-    {
-        public static async Task<StorageFolder> GetBackupFolderAsync(string backupFolderName)
-        {
-            return await FileSystemUtility.GetOrCreateAppFolder(backupFolderName);
-        }
-
-        public static async Task<IReadOnlyList<StorageFile>> GetAllBackupFilesAsync(string backupFolderName)
-        {
-            StorageFolder backupFolder = await GetBackupFolderAsync(backupFolderName);
-            return await backupFolder.GetFilesAsync();
-        }
-      
-        public static async Task<StorageFile> CreateNewFileInBackupFolderAsync(string fileName, CreationCollisionOption collisionOption, string backupFolderName)
-        {
-            StorageFolder backupFolder = await GetBackupFolderAsync(backupFolderName);
-            return await backupFolder.CreateFileAsync(fileName, collisionOption);
-        }
-    }
 }
